Cycle references refactor
This commit introduces the concept of cycle references, along with
several changes that support the feature.

Previously, if a path "cycled", eg A/B/C/A, the paragraphs would be
displayed redundantly as if they were unrelated paragraphs, similar to
how recursive stack frames are not treated differently than others.

However, this is out of sync with what most users expect, as they
understand nodes in the path to be geographic, such that the state of
the session after a cycle is equivalent to before the cycle, as is one's
position before moving in a circle the same as afterwords.

Cycle references also prepare us for the upcoming introduction of
path references, wherein an author will be able to make links to
arbitrary paths of topics. This combined with the removal of cycles by
default will allow several existing concepts to be folded into the
concept of path references; global and import references are just path
references, and redundant parent links are just import self-references
plus the reduction of the resulting cycle; a self-import reference to
the topic subtopic is effectively a back button, etc.

Other changes in this commit are as follows:

`updateView` no longer detects contradictory path and link arguments
but relies on the caller to use correct arguments, unless the
combination is impossible to render in which case it errors.

`updateView` is being replaced in many places by Link#select,
Paragraph#select, or Path.display, as these functions can call
updateView with a consistent path and link.

Text icons are introduced for cycle references, and external link icons
are replaced with text icons in the same style.

Whereas previously encountering a global reference triggered a recursive
call to fetchAndRenderPath, now we just request and render all segments
of the path concurrently from fetchAndRenderPath.

Link#select and Link#execute are introduced to consolidate similar logic
in key and click handlers.

The concept of Link#path is separated out into Link#literalPath,
Link#enclosingPath, Link#inlinePath, and Link#paragraphPath, because now
and going forward these concepts can vary, for example with a cycle link
whose preview path is not its inline path.

If the back button is disabled multiple times we go with the longest
time before showing.

displayPath now introduces animation behavior, wherein if the previous
path shares a prefix with the new path, or if we are going up directly
more than one paragraph, we "animate", ie scroll to the paragraph
that is the prefix or joint before collapsing the previous subpath, and
then update to the new path.

In error 'auto' was used as a scroll behavior instead of 'instant'
in various places.

Similar to what we do with overly large links being put over the top
of the screen, now we detect if scrolling to a link puts the enclosing
paragraph top off the top of the viewport, and if so, we try to prefer
keeping the top of the paragraph in view, unless doing so puts the
link being selected very far down the page. This reduces the amount of
scrolling when iterating between links in a single paragraph.

If the element being scrolled to is off the page, then minDiff is
ignored and scrolling is mandatory.

The concepts of Path#url and Path#rendered are distinguished from one
another in several places.

generateHeader now returns existing header if present.

+++ cli/bulk/file_system_manager.js
@@ -73,7 +73,8 @@ class FileSystemManager {
 
   loadOriginalSelectionFileSet() {
     if (!fs.existsSync('.canopy_bulk_original_selection')) {
-      throw new Error(chalk.red('Expected .canopy_bulk_original_selection file but did not find one'));
+      console.error(chalk.red('Expected .canopy_bulk_original_selection file but did not find one'));
+      return new FileSet({});
     }
     let json = fs.readFileSync('.canopy_bulk_original_selection').toString();
     let selectedFilesList = JSON.parse(json);
@@ -81,7 +82,9 @@ class FileSystemManager {
   }
 
   deleteOriginalSelectionFile() { // this has to be separate from loadOriginalSelectionFileSet in the case where a parsing error prevents processing
-    fs.unlinkSync('.canopy_bulk_original_selection');
+    if (fs.existsSync('.canopy_bulk_original_selection')) {
+      fs.unlinkSync('.canopy_bulk_original_selection');
+    }
   }
 
   getOriginalSelectionFileList() {
